% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/estimate_crowding.R
\name{crowd_no_overlap_ss}
\alias{crowd_no_overlap_ss}
\title{Calculate crowding with genet overlap not allowed}
\usage{
crowd_no_overlap_ss(A, vt, h, alphaG, alphaS, WmatG, WmatS, n_spp, Ctot, Cr,
  b.r, expv, r.U, v.r, size.range)
}
\arguments{
\item{A}{Area of the quadrat.}

\item{h}{}

\item{alphaG}{Alpha parameters (vector with n=n_spp) for growth.}

\item{alphaS}{Alpha parameters (vector with n=n_spp) for survival.}

\item{WmatG}{}

\item{WmatS}{}

\item{n_spp}{Number of species.}

\item{Ctot}{}

\item{Cr}{}

\item{b.r}{}

\item{expv}{}

\item{r.U}{}

\item{v.r}{}

\item{size.range}{}

\item{nt}{Current population vector.}
}
\description{
Calculate crowding with genet overlap not allowed
}
\author{
Andrew Tredennick
}

