% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/estimate_crowding.R
\name{crowd_overlap}
\alias{crowd_overlap}
\title{Calculate crowding with genet overlap allowed}
\usage{
crowd_overlap(A, N, vt, h, alphaG, alphaS, WmatG, WmatS, n_spp, Ctot, Cr, b.r,
  expv, r.U, v.r, v)
}
\arguments{
\item{A}{Area of the quadrat.}

\item{N}{Current population density vector.}

\item{h}{}

\item{alphaG}{Alpha parameters (vector with n=n_spp) for growth.}

\item{alphaS}{Alpha parameters (vector with n=n_spp) for survival.}

\item{WmatG}{}

\item{WmatS}{}

\item{n_spp}{Number of species.}

\item{Ctot}{}

\item{Cr}{}

\item{b.r}{}

\item{expv}{}

\item{r.U}{}

\item{v.r}{}

\item{v}{}

\item{nt}{Current population vector.}

\item{v}{}
}
\description{
Calculate crowding with genet overlap allowed
}
\author{
Andrew Tredennick
}

