% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/make_iteration_matrix.R
\name{expand_W_matrix}
\alias{expand_W_matrix}
\title{Expand crowding matrix}
\usage{
expand_W_matrix(v, u, W)
}
\arguments{
\item{v}{}

\item{u}{}

\item{W}{}
}
\description{
Expand crowding matrix
}
\author{
Andrew Tredennick
}

