% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/demog_stoch_fxns.R, R/make_iteration_matrix.R
\name{GenerateMultivariatePoisson}
\alias{GenerateMultivariatePoisson}
\title{Generate poisson vector}
\usage{
GenerateMultivariatePoisson(pD, samples = 1, R, lambda)

GenerateMultivariatePoisson(pD, samples = 1, R, lambda)
}
\arguments{
\item{pD}{Length of the population vector}

\item{samples}{Number of random samples to generate at each vector element (defaults to 1)}

\item{R}{Population vector covariance matrix from 'get_cov'}

\item{lambda}{Surival*growth contribution (iteration matrix * population vector)}

\item{pD}{Length of the population vector}

\item{samples}{Number of random samples to generate at each vector element (defaults to 1)}

\item{R}{Population vector covariance matrix from 'get_cov'}

\item{lambda}{Surival*growth contribution (iteration matrix * population vector)}
}
\description{
Generate poisson vector

Generate poisson vector
}
\author{
Andrew Tredennick

Andrew Tredennick
}

