% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/demog_stoch_fxns.R, R/make_iteration_matrix.R
\name{get_cov}
\alias{get_cov}
\alias{get_pairs}
\title{Get nt covariance}
\usage{
get_cov(K)

get_cov(K)
}
\arguments{
\item{K}{Combined survival*growth iteration matrix}

\item{K}{Combined survival*growth iteration matrix}
}
\description{
Get nt covariance

Get nt covariance
}
\author{
Andrew Tredennick

Andrew Tredennick
}

