% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/demog_stoch_fxns.R, R/make_iteration_matrix.R
\name{get_pairs}
\alias{get_pairs}
\title{Get pairs}
\usage{
get_pairs(X, pop_vector)

get_pairs(X, pop_vector)
}
\arguments{
\item{X}{}

\item{pop_vector}{}

\item{X}{}

\item{pop_vector}{}
}
\description{
Get pairs

Get pairs
}
\author{
Andrew Tredennick

Andrew Tredennick
}

