% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/make_inits_ipm_ss.R
\name{make_inits_ipm_ss}
\alias{make_inits_ipm_ss}
\title{Function to create iteration matrix size and initial vectors for IPM}
\usage{
make_inits_ipm_ss(n_spp, iter_matrix_dims, max_size)
}
\arguments{
\item{n_spp}{A scalar for the number of species to simulate.}

\item{iter_matrix_dims}{A vector whose length is equal to 'n_spp' and represents
the dimensions of iteration matrix.}

\item{max_size}{A vector whose length is equal to 'n_spp' and represents the
maximum size allowable for a genet from each species.}
}
\value{
Returns a list of matrices and vectors for IPM simulating.
}
\description{
Function to create iteration matrix size and initial vectors for IPM
}
\author{
Andrew Tredennick
}

