% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/make_iteration_matrix.R
\name{make_K_matrix}
\alias{make_K_matrix}
\title{Make iteration matrix based only on mean crowding}
\usage{
make_K_matrix(v, muWG, muWS, rec_params, recs_per_area, growth_params,
  surv_params, do_year, do_spp, h, demo_stoch = FALSE, popv)
}
\arguments{
\item{v}{}

\item{muWG}{}

\item{muWS}{}

\item{rec_params}{}

\item{recs_per_area}{}

\item{growth_params}{}

\item{surv_params}{}

\item{do_year}{}

\item{do_spp}{}

\item{h}{}
}
\description{
Make iteration matrix based only on mean crowding
}
\author{
Andrew Tredennick
}

