% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/make_iteration_matrix_climate.R
\name{make_K_matrix_ss_clim}
\alias{make_K_matrix_ss_clim}
\title{Make iteration matrix based on mean crowding and climate: single species}
\usage{
make_K_matrix_ss_clim(v, muWG, muWS, rec_params, recs_per_area, growth_params,
  surv_params, do_year, do_spp, h, climate)
}
\arguments{
\item{v}{}

\item{muWG}{}

\item{muWS}{}

\item{rec_params}{}

\item{recs_per_area}{}

\item{growth_params}{}

\item{surv_params}{}

\item{do_year}{}

\item{do_spp}{}

\item{h}{}
}
\description{
Make iteration matrix based on mean crowding and climate: single species
}
\author{
Andrew Tredennick
}

