% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/make_iteration_matrix.R
\name{make_K_values}
\alias{make_K_matrix}
\alias{make_K_values}
\title{Make iteration matrix}
\usage{
make_K_values(v, u, muWG, muWS, rec_params, recs_per_area, growth_params,
  surv_params, do_year, do_spp)
}
\arguments{
\item{v}{}

\item{u}{}

\item{muWG}{}

\item{muWS}{}

\item{rec_params}{}

\item{recs_per_area}{}

\item{growth_params}{}

\item{surv_params}{}

\item{do_year}{}

\item{do_spp}{}

\item{demographic_stochasticity}{}
}
\description{
Make iteration matrix
}
\author{
Andrew Tredennick
}

