% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/make_iteration_matrix_climate.R
\name{make_K_values_clim}
\alias{make_K_values_clim}
\title{Make iteration values}
\usage{
make_K_values_clim(v, u, muWG, muWS, rec_params, recs_per_area, growth_params,
  surv_params, do_year, do_spp, climate)
}
\arguments{
\item{v}{}

\item{u}{}

\item{muWG}{}

\item{muWS}{}

\item{rec_params}{}

\item{recs_per_area}{}

\item{growth_params}{}

\item{surv_params}{}

\item{do_year}{}

\item{do_spp}{}
}
\description{
Make iteration values
}
\author{
Andrew Tredennick
}

