% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/util_functions.R
\name{sum_cover}
\alias{sum_cover}
\title{Function to sum total cover of each species}
\usage{
sum_cover(v, nt, h, A)
}
\arguments{
\item{v}{}

\item{nt}{The population vector.}

\item{h}{}

\item{A}{Area of the quadrat (scalar)}
}
\description{
Function to sum total cover of each species
}
\author{
Andrew Tredennick
}

