% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/util_functions.R
\name{sum_N}
\alias{sum_N}
\title{Function to sum total density of each species}
\usage{
sum_N(nt, h)
}
\arguments{
\item{nt}{The population vector.}

\item{h}{}
}
\description{
Function to sum total density of each species
}
\author{
Andrew Tredennick
}

