% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/util_functions.R
\name{varN}
\alias{varN}
\alias{var_N}
\title{Function to calculate size variance of each species}
\usage{
varN(v, nt, h, Xbar, N)
}
\arguments{
\item{v}{}

\item{nt}{The population vector.}

\item{h}{}

\item{Xbar}{}

\item{N}{Density vector for each size class.}
}
\description{
Function to calculate size variance of each species
}
\author{
Andrew Tredennick
}

