% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_crowding.R
\name{wrijG}
\alias{get_crowd_growth}
\alias{wrijG}
\title{Function to calculate size-dependent crowding for growth, assuming no overlap}
\usage{
wrijG(r, i, j, r.U, Cr, Ctot)
}
\arguments{
\item{r}{}

\item{i}{}

\item{j}{}

\item{r.U}{}

\item{Cr}{}

\item{Ctot}{}
}
\description{
Function to calculate size-dependent crowding for growth, assuming no overlap
}

