% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_crowding_ss.R
\name{wrijG_ss}
\alias{get_crowd_growth_ss}
\alias{wrijG_ss}
\title{Function to calculate size-dependent crowding for growth, assuming no overlap}
\usage{
wrijG_ss(r, i, r.U, Cr, Ctot)
}
\arguments{
\item{r}{}

\item{i}{}

\item{r.U}{}

\item{Cr}{}

\item{Ctot}{}

\item{j}{}
}
\description{
Function to calculate size-dependent crowding for growth, assuming no overlap
}

